refactor(config): separate zod schema and parsing logic for better maintainability and scalability
